Refactoring UI into individual TypeScript files.
